refactor: rename AngularJS MetricChooser
so we can introduce Angular MetricChooser at its place

ref #2318
