+++ visualization/app/codeCharta/ui/edgeChooser/edgeChooser.module.ts
@@ -2,7 +2,7 @@ import "../../state/state.module"
 import angular from "angular"
 import { edgeChooserComponent } from "./edgeChooser.component"
 import { downgradeComponent } from "@angular/upgrade/static"
-import { MetricTypeHoveredComponent } from "../metricChooser/metricTypeHovered/metricTypeHovered.component"
+import { MetricTypeHoveredComponent } from "../metricChooserDeprecated/metricTypeHovered/metricTypeHovered.component"
 
 angular
 	.module("app.codeCharta.ui.edgeChooser", ["app.codeCharta.state"])
