+++ visualization/app/codeCharta/ui/metricChooserDeprecated/metricChooser.e2e.ts
