+++ visualization/app/codeCharta/ui/metricChooserDeprecated/metricTypeHovered/isHoveredNodeALeaf.selector.ts
