+++ visualization/app/codeCharta/ui/metricChooserDeprecated/metricValueHovered/metricValueHovered.component.ts
