+++ visualization/app/codeCharta/ui/ui.ts
@@ -12,7 +12,7 @@ import "./codeMap/codeMap.module"
 import "./colorSettingsPanel/colorSettingsPanel.module"
 import "./dialog/dialog.module"
 import "./fileExtensionBar/fileExtensionBar.module"
-import "./metricChooser/metricChooser.module"
+import "./metricChooserDeprecated/metricChooser.module"
 import "./resetSettingsButton/resetSettingsButton.module"
 import "./filePanel/filePanel.module"
 import "./ribbonBar/ribbonBar.module"
